backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-json-logger==2.0.7
websockets==12.0
httpx==0.25.2
pydantic==2.5.0
bcrypt==4.1.2
pyserial==3.5
requests==2.31.0
# SQLite is included with Python standard library